Fixed clippy warnings

+++ src/dense/linalg/lapack/inverse.rs
@@ -22,7 +22,7 @@ where
         let n = self.shape()[1];
         assert_eq!(m, n, "Matrix must be square for inversion.");
         let mut out = DynArray::new_from(self);
-        let mut ipiv = vec![0 as i32; m];
+        let mut ipiv = vec![0_i32; m];
 
         Item::getrf(m, n, out.data_mut(), m, &mut ipiv)?;
         Item::getri(n, out.data_mut(), m, &ipiv)?;
